[solver] Hoist one-shot subprocess helpers out of bitwuzllob
Move the verdict scanner, formula-path/model-prog option plumbing and the
fork/exec/verdict-scan runner from bitwuzllob_conv.cpp into a shared
oneshot_process namespace in the smtlib library, parameterized by a backend
name, so further one-shot subprocess backends can reuse them. Behaviour is
unchanged; parse_verdict_line gains a unit test now that it is exported.
